Followup to 3f40ce7: move defensive copy into get_download_url itself
Summary:
This way, callers don't have to worry about it.

While I was in there, I tried to make get_download_url a bit more
threadsafe. I'm still not positive whether it's commonly going to be
invoked from multiple threads (and how the request_iterator interacts
with the threadpool), but this code seems at least as likely as any
other to be used that way.

Ensure that the download URL, headers, and expiry time are all packed
into a single tuple rather than as 3 fields. This ensures that two
racing calls to get_download_url can't mix-and-match their values: one
or the other wins for all of the fields. Also, once we write those
values to "global" (really, instance) variables, we should assume they
could be changed out from under us. So don't use fields in "self" as
intermediate variables in our method; only write to them at the end.

Test Plan: dxpy integration tests

Reviewers: psung

Differential Revision: https://dnanexus.phacility.com/D758